Authentication and Permissions

+++ controllers/bootcamps.js
@@ -29,26 +29,40 @@ exports.getBootcamp = asyncHandler(async (req, res, next) => {
 // @route   POST /api/v1/bootcamps
 // @access  Private  
 exports.createBootcamp = asyncHandler(async (req, res, next) => {
-        const bootcamp = await Bootcamp.create(req.body);
+    req.body.user = req.user.id;
 
-        res.status(201).json({
-            success: true, 
-            data: bootcamp
-        });
+    const publishedBootcamp = await Bootcamp.findOne( {user: req.user.id });
+
+    if(publishedBootcamp && req.user.role != 'admin'){
+        return next(new ErrorResponse(`The user with ID ${req.user.id} has already published a bootcamp`, 404))
+    }
+
+    const bootcamp = await Bootcamp.create(req.body);
+
+    res.status(201).json({
+        success: true, 
+        data: bootcamp
+    });
 })
 
 // @desc    Update new bootcamp
 // @route   PUT /api/v1/bootcamps/:id
 // @access  Private  
 exports.updateBootcamp = asyncHandler(async (req, res, next) => {
-        const bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
-            new : true,
-            runValidators: true
-        });
+        let bootcamp = await Bootcamp.findById(req.params.id); 
     
         if (!bootcamp) {
             return next(new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`,404));
         }
+
+        if(bootcamp.user.toString() != req.user.id && req.user.role !== 'admin'){
+            return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`,401));
+        }
+
+        bootcamp = await Bootcamp.findByIdAndUpdate(req.params.id, req.body, {
+            new: true,
+            runValidators: true
+        });
     
         res.status(200).json({success: true, data: bootcamp });
 })
@@ -63,6 +77,10 @@ exports.deleteBootcamp = asyncHandler(async (req, res, next) => {
         return next(
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`,404)
         );
+    }
+
+    if(bootcamp.user.toString() != req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.params.id} is not authorized to delete this bootcamp`,401));
     }
 
     await bootcamp.deleteOne();
@@ -104,6 +122,10 @@ exports.bootcampPhotoUpload = asyncHandler(async (req, res, next) => {
         return next(
             new ErrorResponse(`Bootcamp not found with id of ${req.params.id}`,404)
         );
+    }
+
+    if(bootcamp.user.toString() != req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.params.id} is not authorized to update this bootcamp`,401));
     }
 
     if (!req.files) {
